#6 ref: AppBar와 Hamburger 메뉴를 새로운 파일로 분리
